# src/main/java/inter/app/model/SingleDigit.java
package inter.app.model;

public class SingleDigit {
    private int input;
    private int base;
    private int digit;

    public void setInput(int input) {
        this.input = input;
    }

    public void setBase(int base) {
        this.base = base;
    }

    public void setDigit(int digit) {
        this.digit = digit;
    }

    public int getInput() {
        return input;
    }

    public int getBase() {
        return base;
    }

    public int getDigit() {
        return digit;
    }
}